Review corrections
 -Fix compiler warnings by typecast
 -Add missing brackets to few places
 -Remove additional line change

+++ library/ssl_cli.c
@@ -700,7 +700,7 @@ static int ssl_generate_random( mbedtls_ssl_context *ssl )
 
 #if defined(MBEDTLS_HAVE_TIME)
     t = mbedtls_time( NULL );
-    p = mbedtls_platform_put_uint32_be( p, t );
+    p = mbedtls_platform_put_uint32_be( p, (uint32_t) t );
     MBEDTLS_SSL_DEBUG_MSG( 3, ( "client hello, current time: %lu", t ) );
 #else
     if( ( ret = mbedtls_ssl_conf_get_frng( ssl->conf )
